requirements.txt
boto3==1.34.114
botocore==1.34.114
et-xmlfile==1.1.0
jmespath==1.0.1
openpyxl==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
s3transfer==0.10.1
six==1.16.0
urllib3==2.2.1
